refactor(cli): extract class resolution and model-type derivation into cli.resolution
`derive_model_type` is shared: the eval session picks a model class from it, and
recipe resolution picks a capability layer. Sharing it left `cli/infer/recipe.py`
importing `cli/leaderboard/session.py` — a new sideways edge between the two CLI
subpackages, for a function that has nothing to do with running a session.

The class-resolution cluster it sits on moves with it into `sieval/cli/resolution.py`:
`load_class_from_path`, `load_class_from_name`, `resolve_class`,
`resolve_dataset_class`, `resolve_task_class`, `_is_missing_module_error`,
`_guess_submodule_names`, `derive_model_type`, `validate_named_config_map`, plus
`DATASET_MODULE` / `TASK_MODULE`. Bodies are byte-identical — sliced by line
number, not retyped. The new module imports nothing from sieval (it reaches
`sieval.tasks` / `sieval.datasets` through `importlib` at call time), so both
subpackages can depend on it as a leaf.

This is about the dependency graph, not load time: `sieval/cli/__init__.py`
builds the whole app, so importing any `sieval.cli` submodule already pulls in
the session either way. `sieval/cli/infer/*.py` now contains no reference to
`cli.leaderboard`.

`_validate_named_config_map` is promoted to `validate_named_config_map`: it is
read from two modules now, and the import policy treats a cross-module `_name`
in production code as a smell to fix rather than to spell out.

Import sites updated: `session` (which keeps `RootConfigDict` and everything
session-specific), `validation`, `run`, `infer/recipe`. `session` no longer
re-exports these — `CLAUDE.md`'s import policy names `sieval.cli.resolution` as
the canonical path so it does not drift back.

Tests move to `tests/unit/cli/test_resolution.py` to keep the mirror rule, along
with the `_project_sieval_first` sys.path fixture only they use. Four `patch()`
targets are repointed to the new module — `derive_model_type` and
`resolve_task_class` resolve names through their own module globals, so patching
`session`'s copy would silently no-op. The `_setup_tasks` / `_setup_datasets`
patches stay on `session`, which still resolves those names through its globals.
Suite count is unchanged at 3471, so nothing was dropped in the move.
